move quotes into own component
